+++ core/core.pl
@@ -10261,6 +10261,11 @@ create_logtalk_flag(Flag, Value, Options) :-
 
 % use_module/1 entity directive
 
+'$lgt_compile_logtalk_directive'(use_module(Module), Ctx) :-
+	'$lgt_pp_module_'(_),
+	% compiling a module as an object
+	throw(error(domain_error(directive, use_module/1), directive(use_module(Module)))).
+
 '$lgt_compile_logtalk_directive'(use_module(Aliases), Ctx) :-
 	'$lgt_prolog_feature'(modules, unsupported),
 	\+ '$lgt_pp_module_'(_),